revert to static class
